Move LSPS2 service event handling into `liquidity/service/lsps2.rs`
